Move triggered functions to a separate directory
