+++ functions/src/triggered/aws-poller.ts
